Version 1.954

+++ events/message.js
@@ -9,63 +9,28 @@ exports.run = async (Bot, Message) => {
     return;
   }
 
-  if (Bot.isURL(Message.content)){
-    Message.delete();
+  const AntiURL = await Bot.Database.get(`ServerData_${Message.guild.id}.AntiURL`)
 
-    return Message.reply("Whoa there! You cannot send links here.").then(m => m.delete({ timeout: 5000 }))
+  if (AntiURL && AntiURL === "on" && Bot.isURL(Message.content)){
+    try {
+      Message.delete();
+    } catch (err){
+      Message.channel.send(`${Message.author} sent a url, but I cannot delete it. Please give me permision to delete messages.`).then(m => m.delete({ timeout: 1000 }))
+    }
+    
+    return Message.reply("you cannot send links here.").then(m => m.delete({ timeout: 1000 }))
   }
 
   const AntiSpam = await Bot.Database.get(`ServerData_${Message.guild.id}.AntiSpam`)
-  const AntiSpamMap = Bot.AntiSpamMap
 
   if (AntiSpam && AntiSpam === "on"){
-    const Limit = 5
-    const Time = 12
-    const Diff = 3
-
-    if (AntiSpamMap.has(Message.author.id)) {
-      const UserData = AntiSpamMap.get(Message.author.id)
-      const { LastMessage, Timer } = UserData
-      var MessageCount = UserData.MessageCount
-      const Difference = Message.createdTimestamp - LastMessage.createdTimestamp
-
-      if (Difference > Diff * 1000) {
-        clearTimeout(Timer)
-
-        UserData.MessageCount = 1
-        UserData.LastMessage = Message
-        UserData.LastMessage = Message
-        UserData.Timer = setTimeout(() => {
-          AntiSpamMap.delete(Message.author.id)
-        }, Time * 1000)
-        AntiSpamMap.set(Message.author.id, UserData)
-      } else {
-        ++MessageCount
-
-        if (parseInt(MessageCount) === Limit) {
-          Message.delete({ timeout: 5 })
-          Message.reply(`please stop spamming! You're now on a cooldown of ${Time} seconds.`).then(m => m.delete({ timeout: 5000 }))
-        } else {
-          UserData.MessageCount = MessageCount
-          AntiSpamMap.set(Message.author.id, UserData)
-        }
-      }
-    } else {
-      let Timeout = setTimeout(() => {
-        AntiSpamMap.delete(Message.author.id)
-      }, Time * 1000)
-      AntiSpamMap.set(Message.author.id, {
-        MessageCount: 1,
-        LastMessage: Message,
-        Timer: Timeout
-      })
-    }
+    Bot.AntiSpam.message(Message)
   }
 
   const Prefix = await Bot.Database.get(`ServerData_${Message.guild.id}.Prefix`)
 
   if (Prefix) {
-    if (!Message.content.startsWith(Prefix)) {
+    if (!Message.content.startsWith(Prefix)){
       return
     }
   } else {
